Combine EVarDef with EDef.
This removes constants, which aren't needed in this language design because it's designed for scripting, not large applications. It may be somewhat error prone though, since you have to use the '<-' operator to reassign variables, instead of '='. If you forget and use '=' instead, it'll silently create a whole new variable.